Revert "refactor: common domain 패키지 생성"
This reverts commit 7148cb0b63dfc1165cdecc1e1c51e28cb810221a.
