+++ src/main/java/com/prography/lighton/member/domain/entity/association/PreferredArtist.java
@@ -2,12 +2,9 @@ package com.prography.lighton.member.domain.entity.association;
 
 
 import com.prography.lighton.artist.domain.entity.Artist;
-import com.prography.lighton.common.domain.BaseEntity;
+import com.prography.lighton.common.BaseEntity;
 import com.prography.lighton.member.domain.entity.Member;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 
 @Entity
 public class PreferredArtist extends BaseEntity {
