+++ src/main/java/com/prography/lighton/performance/domain/entity/vo/Payment.java
@@ -1,6 +1,12 @@
 package com.prography.lighton.performance.domain.entity.vo;
 
-import jakarta.persistence.Embeddable;
+import com.prography.lighton.common.BaseEntity;
+import com.prography.lighton.performance.domain.entity.enums.Seat;
+import jakarta.persistence.*;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.SQLRestriction;
+import lombok.AllArgsConstructor;
 
 @Embeddable
 public class Payment {
